refactor(auth): take out AsyncStoraged and some simple changes
AsyncStoraged create a bug when used with UseEfect. The information of the user, token and refresh
tokens will be saved only in the useContext. A reset funcion was created in the Login screen and to
reset all when the login was fall. The patient status change the collor: Red to infected